+++ packages/validate/CHANGELOG.md
@@ -1,5 +1,5 @@
 # 3.0.0-alpha.1
-* Add `MapField`.
+* Add `MapField` and `UriField`.
 
 # 3.0.0-alpha
 * Rewrite, based on `Field` and `Form`, rather than strings.